src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/content-reprojection-fallback-crash-expected.txt

CONSOLE ERROR: line 13: Uncaught TypeError: host1.createShadowRoot is not a function
A crash should not happen when fallback elements are reprojected.

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
